Expanded MANOVA testing
test_nonlinear_oned checks the performance of the MANOVA test on nonlinear data in one dimension. Generates two sets of data with a nonlinear relationship, runs the MANOVA test on them.
test_zero_variance: Checks that the MANOVA test raises an error when one of the input datasets has zero variance.